+++ src/mealie/__init__.py
@@ -1,10 +1,13 @@
 from .categories import CategoriesMixin
 from .client import MealieClient
+from .foods import FoodsMixin
 from .group import GroupMixin
 from .mealplan import MealplanMixin
 from .recipe import RecipeMixin
 from .shopping_list import ShoppingListMixin
 from .tags import TagsMixin
+from .tools import ToolsMixin
+from .units import UnitsMixin
 from .user import UserMixin
 
 
@@ -12,6 +15,9 @@ class MealieFetcher(
     RecipeMixin,
     CategoriesMixin,
     TagsMixin,
+    FoodsMixin,
+    UnitsMixin,
+    ToolsMixin,
     ShoppingListMixin,
     MealplanMixin,
     UserMixin,